chore: Move generator code to generators folder
