+++ packages/graphback/src/generators/resolvers/index.ts
@@ -1,4 +1,4 @@
-import { Type } from '../ContextTypes'
+import { Type } from '../../ContextTypes'
 import { buildResolverTargetContext, createCustomContext, TypeContext } from './knex/targetResolverContext';
 import { generateGraphbackResolvers } from './outputResolvers/resolverTemplate';
