+++ packages/graphback/src/generators/schema/index.ts
@@ -1,4 +1,4 @@
-import { Type } from '../ContextTypes';
+import { Type } from '../../ContextTypes';
 import { generateSchema } from './schemaTemplate';
 import { buildTargetContext, createCustomSchemaContext, TargetContext } from './targetSchemaContext';
